fix3 task0

+++ 0x14-bit_manipulation/0-binary_to_uint.c
@@ -19,7 +19,7 @@ return (1 + len_string(str + 1));
 unsigned int binary_to_uint(const char *b)
 {
 int nb_digit = 0;
-unsigned int  sum = 0;
+unsigned int long  sum = 0;
 unsigned long int factor = 0;
 int i = 0;
